chore: clean up single-used function
