+++ packages/state-transition/src/util/validator.ts
@@ -84,14 +84,6 @@ export function getMaxEffectiveBalance(withdrawalCredentials: Uint8Array): numbe
   }
 }
 
-export function getActiveBalance(state: CachedBeaconStateElectra, validatorIndex: ValidatorIndex): number {
-  const validatorMaxEffectiveBalance = getMaxEffectiveBalance(
-    state.validators.getReadonly(validatorIndex).withdrawalCredentials
-  );
-
-  return Math.min(state.balances.get(validatorIndex), validatorMaxEffectiveBalance);
-}
-
 export function getPendingBalanceToWithdraw(state: CachedBeaconStateElectra, validatorIndex: ValidatorIndex): number {
   return state.pendingPartialWithdrawals
     .getAllReadonly()
